Fix config_facts.py not storing data in certain cases
In some cases, depending on the order in which dictionary items get
processed, some data that belongs to a single-level key can end up not
actually being copied into `data` (the formatted dictionary that's then
used by other test scripts). This was actually happening with the
VLAN_INTERFACE table where `proxy_arp` was present in sonic-cfggen, but
was silently being dropped here.

Fix this issue, and fix a test script that would get broken because
there are non-IP addresses present now.

This was seen with Python 3, where there is a different order in which
dictionary items ended up getting processed.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>