[sanity] Restart BGPd in neighbor VMs as part of adaptive recovery (#4060)
Adaptive recovery does not restore BGM on neighbor hosts completely. BGP router bgp no shutdown is not enough. If BGPd is killed, it needs to be restarted inside EOS VMs to establish BGP with the DUT. Neighbor VMs can be left in bad state (due to test or human error), and sanity recovery path should be able to restore the testbed health.

=> This change restores BGPd inside neighbor VMs as part of adaptive recovery process. This would prevent test failures due to testbed being unhealthy.