Adding a few more fixes to the tests
Details: basing interfaces list on config_db (which is done by get_port_map)
should be done in all cases not only where asic is not None.
Also, added get_dev_conn to test_check_sfp_using_ethtool, as it is also used in that way
in other sfp tests.