Disable syslog rate limit to only disable on containers that are in the critical services list
On a supervisor card in a VoQ chassis, the number of syncd/teamd/swss containers
depends on the number of Switch Fabric cards present. The dockers that are created
are present in the critical_services_list. So, check if the docker (service) being
disabled is in the critical_service_list before trying to disable syslog rate
limit on that docker (service).