collect_dut_info to check if we have front-end and/or back-end asics
PR# 3696 added container names in meta data file for frontend and backend ASIC on which autorestart test will be performed for multi-asic platform.
This assumes that on a multi-asic platform, there are always front-end and back-end asics.

However, on a VoQ chassis
  - linecards don't have back-end asics.
  - supervisor card has niether front-end nor back-end asics

So, need to check if we have front-end or back-end asics before trying to add them to asic_services.