Fix T2 fib_info for routes present a multiple asics on the same linecard.
In a T2 topology, the same route being advertised on multiple asics.
In this case, the route entry would have both the eBGP via front-panel port and the inband port as nexthops.
When generating the t2 fib, we were skipping routes that had inband port as nexthop (The nexthop port in 'Int'). Thus, these
routes were getting discarded in the fib_info.

Fix is not check if there are any other 'Ext' nexthops, and if so, don't skip this route.