[test_hash]: Clarify negative test for hash key 'ingress-port' in comments (#4064)
The 'ingress-port' key is not used in hash of FIB forwarding. If the 'ingress-port'
key is included in HASH_KEYS of the FIB hash testing, a negative test will be run.
The PTF script will try to inject same packet to different ingress ports and expect
that they are forwarded from same egress port. This change added comments
in test scripts to clarify this to avoid confusion.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>