[test_sub_port_interfaces] Skip subport test on unsupported platform
What is the motivation for this PR?
SAI reports SAI_STATUS_NOT_SUPPORTED error while config the support interface on some platforms. Need to mark it to skip on  those platforms.

How did you do it?
Use a fixture to check if need to skip or not.

How did you verify/test it?
Run test_sub_port_interfaces.py, it should be skipped on those platforms.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>