Remove pretest label from test_nbr_health (#4116)
In nightly regression the test test_nbr_health runs as part of pretest list of cases. However, there is no sanity check in this test. This leads to pretest failure if testbed is unhealthy.
Additionally, sanity can't be enabled for this test as it is also run in public KVM testbed deployment, where deploy-mg is done after this testcase is executed:
https://github.com/Azure/sonic-jenkins/blob/master/azure-pipeline/provkvmtest.sh#L367

Remove pretest label, and run test_nbr_health as a standalone testcase.