Refactor add_fail_step_to_reboot fixture to support upgrade path test (#4123)
The add_fail_step_to_reboot fixture checks the reboot type based on the testcase name. This works for test_warm_reboot and test_fast_reboot. But, this fails for test_upgrade_path.
There should be a better way to find reboot type in the caller testcase.
How did you do it?
Move the logic inside add_fail_step_to_reboot to a function inside the fixture. The fixture yields this function.
Reboot type is passed as an argument to the yielded function.