[acl] Run garp service just for dual_tor
What is the motivation for this PR?
After running garp service, DUT will flood the arp packets to all the other ports under the same vlan interface, which means ptf needs to poll N x N packets in one second to do the verification. (N is the number of ports in the vlan interfaces on DUT). So even ptf has already received the expected packet, ptf may miss doing the verification on the DUT which has lots of ports. As this garp service is just required on dual-tor, so we can skip it on single-tor.

How did you do it?
Check the typo name, just import garp service for dual-tor. For the next
step improvement, we may need to consider to increase the interval of
garp service.

How did you verify/test it?
Run test_acl.py on 7260 which has 112 ports in the same vlan interface.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>