Optimize vm_topology to reduce time required for add-topo
Currently, to add a topology using tool 'testbed-cli.sh add-topo', most of the time is spent
in calling the customized vm_topology module to bind interfaces and bridges for PTF docker
and neighbor VMs. Before apply any operation on an interface, usually we need to firstly
check existence of the interface in host or in docker. The vm_topology module runs command
"ifocnfig -a" and parse the output to find out all the interfaces. Then based on this
information to check existence of interface. On a test server with multiple testbeds, there
could be hundereds of interfaces and bridges in the output of "ifconfig -a". Even worse,
checking existence of interface is a frequent operation. So, it could take more than 20
minutes to run the vm_topology module.

This change mainly improved the way of checking existence of interface in the vm_topology
module. Instead of run "ifconfig -a" every time, the new way is to run "ifconfig <intf_name>"
which is much quicker. With this change, the vm_topology needs around 100 seconds for
t1-lag topo and around 40 seconds for t0 topology.

Other improvements made in this PR:
* Add timestamp to log generated by vm_topology.
* The ceos_network module is also improved in similar way.
* Skip exporting the PTF docker if vm_type is not "vsonic".
* Remove the bridges for CEOS during 'testbed-cli.sh remove-topo' if vm_type is "ceos".

Signed-off-by: Xin Wang <xiwang5@microsoft.com>