Make "ip neigh flush $ptf_ip" best effort in testbed-cli.sh add-topo
The add_topo function in testbed-cli.sh runs "ip neighbor flush $ptf_ip" to delete
obsoleted arp entry for the PTF IP to ensure that the newly created PTF IP is
reachable immediately. However, this command may fail with error message
`Failed to send flush request: Operation not permitted` ocassionally. This would
cause none-zero RC for the 'testbed-cli.sh add-topo' task. Since this command
is a best effort operation, so it would be better for it to always return RC 0.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>