Merged PR 133: Add pipeline yaml for redeploy topology of a testbed
This pipeline yaml is intended as a replacement of the Jenkins switch-topo-branch job. However, the yaml is not an exact replication of the original job. Purpose of the Jenkins job is to switch topology of a testbed between 201811 and later branch. This is not needed in nowadays. So, I simply created the pipeline as a tool for topology redeployment of specified testbed.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>