Fix the OVS bridges not removed issue in cleanup.sh
During 'testbed-cli.sh add-topo', a tool 'cleanup.sh' is generated to test server.
This script is for cleaning up  all the VMs, bridges and dockers created during
add-topo. However, the script can't really remove all the OVS bridges as it
intended to. This change fix the issue. With this fix, this script is able to cleanup
all the OVS bridges as well.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>