[snmp] address random shared memory check failure (#4137)
What is the motivation for this PR?
test_snmp_memory.py is randomly failing.

How did you do it?
Shared memory is a changing number on target device, therefore there is a chance that the snmp and system read took different readings by timing.

This is an enhancement but I don't think this change addresses the issue fully. The reason for these random failures was mainly because the time we collected the snmp facts and the time we read these information from the dut would be different, So memory allocation / usage changes over time. Anything dynamic could see a jump of more than 4%. 4% is too artificial to work on all cases.

How did you verify/test it?
Run snmp test.

Signed-off-by: Ying Xie ying.xie@microsoft.com