Fix the first testbed missing in server recover summary log issue
The recover_server.py tool prints a summary log reporting recovery
results of each testbed. When vm type is "ceos", the first testbed
is missing from the recover summary log. The issue was introduced
after the recover_server.py script was enhanced to support ceos.
The code printing summary needs to skip the job for "start-vms".
When vm type is ceos, the start-vms job is not needed. However,
the code printing summary always assume that a start-vms job
is there and unnecessary skipped a job when vm type is ceos.
Consequently, the job for the first testbed is skipped.

The fix is to only skip log of the "start-vms" job when vm type is
not ceos.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>