Fix ansible variable VM_targets undefined issue for ptf topology
PR #4156 replaced variable VM_hosts with VM_targets when call the
vm_topology module. With this change, the vm_topology will create
bridges for VMs used by current testbed instead of create bridges
for all VMs available on the test server.
However, the change missed the ptf topology scenario. In case of
ptf topology, the VM_targets ansible varialbe is not defined.

The fix is to assign default value `[]` to VM_targets when it is not
defined.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>