add mac for ping dut and ping from servers (#4152)
A part of the ping packets for advanced reboot is not provided with a source mac address. It follows that the default mac address (00:06:07:08:09:0A) is used for packets from different ports. The DUT may generate a large number of unexpected fdb notifications. Assign mac address to ping packets to avoid the behavior.