Fix the OVS bridges not removed issue in cleanup.sh (#4202)
What is the motivation for this PR?
During 'testbed-cli.sh add-topo', a tool 'cleanup.sh' is generated to test server.
This script is for cleaning up all the VMs, bridges and dockers created during
add-topo. However, the script can't really remove all the OVS bridges as it
intended to.

How did you do it?
This change fixed the command for removing OVS bridges.
With this fix, this script is able to cleanup all the OVS bridges as well.

How did you verify/test it?
Run this script on a test server. Then check OVS bridges.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>