support multi type of sai test container
use a parameter specify the container type, then it can download, deploy
and restart the specified sai test container, right now the container
can be saiserver and syncd.

For saiserver container it is a standard container, which just expose
the sai local interface to RPC APIs, which support remote invocation.
For Syncd sai test container, it is using a syncd-rpc container, with
all the syncd container environment, like naming and starting script.
After testing, the environment will be recovered.