Fix ansible variable VM_targets undefined issue for ptf topology (#4229)
What is the motivation for this PR?
PR #4156 replaced variable VM_hosts with VM_targets when call the
vm_topology module. With this change, the vm_topology module will
create bridges for VMs used by current testbed instead of create bridges
for all VMs available on the test server.

However, the change missed the ptf topology scenario. In case of
ptf topology, the VM_targets ansible variable is not defined. Calling
the vm_topology module will fail.

How did you do it?
The fix is to assign default value [] to VM_targets when it is not defined.

How did you verify/test it?
Test add-topo/remove-topo for ptf and t0 topologies.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>