Merged PR 147: Improve the recover_server pipeline to support archiving logs
The recover_server pipeline generates some logs under /tmp folder of the agent running the job. It is tedious to check those logs on agent for debugging.

This change added steps to collect and archive the logs. With this change, we can check the build artifacts to view the logs.