[dualtor][ipinip] Increase the interval of garp service
What is the motivation for this PR?
Garp packets will flood to all interfaces in the same vlan,  so increase the packet interval of garp could mitigate this potential issue for the cases which requires garp service to .

How did you do it?
Increase the interval from 1s to 10s. It is supposed to work well as 10s is still much less than the default age time of arp and fdb.

How did you verify/test it?
Run the garp service.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>