Add a Probe test to checking the test environment
As the sai test container might need a while to get ready, we need to
make sure the test container is ready before we can run the actual test.
Then we need to wait for a while and then check if the sai test
container with some simply test case. Then we can make sure the test
container is running.