[dualtor][test_orchagent_mac_move] Modify the test address to avoid confliction with garp service (#4255)
What is the motivation for this PR?
Mac move test case currently uses 192.168.0.100 as the neighbor address. But Garp service also uses the address from 192.168.0.2 for each interface. So there is a confliction issue on the platform which has more than 98 downlink ports.

How did you do it?
Change the address to 192.168.0.250.

How did you verify/test it?
Run the mac move test.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>