[dualtor][ipinip] Increase the interval of garp service (#4239)
* [dualtor][ipinip] Stop garp service for single tor

What is the motivation for this PR?
Garp packets will flood to all interfaces in the same VLAN,
even ptf has already received the expected packet, ptf may miss doing the verification on the DUT which has lots of ports. As this garp service is just required on dual-tor, so we can skip it on single-tor.

How did you do it?
Stop garp service on single tor
Use verify_packet to take replace verify_packet_any_port to improve performance.

How did you verify/test it?
Run the test_ipinip.py.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>

* [dualtor][ipinip] Increase the interval of garp service

What is the motivation for this PR?
Garp packets will flood to all interfaces in the same vlan,  so increase the packet interval of garp could mitigate this potential issue for the cases which requires garp service to .

How did you do it?
Increase the interval from 1s to 10s. It is supposed to work well as 10s is still much less than the default age time of arp and fdb.

How did you verify/test it?
Run the garp service.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>