Modified VOQ portchannel testcases to get asic with portchannels configured. (#3754)
Modified existing voq portchannel testcases to get ASIC with portchannel configured in case of multi-asic.

Summary:
In case of multi-asic testcase was skipping testcase beacuse it was not able to get ASIC with portchannel configured. Modified testcase to ger ASIC with pc configured else skip test testcase if portchannel not configured.