Ssh stress test (#3900)
Description of PR
Summary:
This test makes sure that the DUT can perform even when under stress of multiple concurrent commands being run in parallel.

Type of change
Approach
What is the motivation for this PR?
SSH stress test is present on Arista test suite, but not in SONiC. This test aims to put stress on the DUT by opening a number of parallelized SSH connections simultaneously running commands.

This test also reports the number of consecutive SSH connections that can be made to the DUT without severely reducing DUT operation (max 20).

How did you do it?
Commands are run in parallel, CPU and Memory usage are tracked on the DUT. The test expects that the memory should drop after the commands finish running and the SSH connections are terminated. The runtime for each command is also compared to a baseline taken for the command before the tests starts.

The number of maximum SSH connections is determined by adding SSH connections until a command times out. This number is reported.

How did you verify/test it?
Test passed locally. The test verifies all copied files are removed from the DUT.

Co-authored-by: Andoni Sanguesa <t-asanguesa@microsoft.com>