Ignoring ip protocol 253 in IPv6 packets in the fib hash test. (#4274)
What is the motivation for this PR?
When a DNX chipset gets an IPv6 packet with ip protocol 253, it strips out 4 bytes from its payload
before forwarding the packet out. This issue was raised against BCM

https://github.com/Broadcom-Switch-Collab/Zeus_SAI/issues/100

How did you do it?
Recommendation of BCM and MSFT is to avoid ip protocol 253 (similar to protocol 2 and 254) in the fib hash test

How did you verify/test it?
Ran fib_hash test against linecards in a VoQ Chassis with DNX chipset.