[test][nhop resources] Ignore NHOP resource count check on Mellanox (#4275)
Fix Azure/sonic-buildimage#8727

The CRM resources in Mellanox seem to be shared. The test creates NHOP groups as advertised (up to a max of 1024). On Mellanox even after creating all the NHOP groups, NHOP resources are not '0'.

The primary purpose of this test is to make sure SONiC doesn't crash when NHOP groups beyond the advertised number is created. That check is not skipped.