Add back retry for running commands in the vm_topology module
PR #4156 optimized the performance of the vm_topology module. However, the retry
logci of running a command is also removed. On some testbeds, command like "ifconfig -a"
may fail ocassionally and need to retry. Without the retry logic, add-topo and
restart-ptf may fail ocassionally.

This change added back the retry logic of running commands in the vm_topology module and
should can increase robustness of calling the vm_topology module.
The ceos_network module is also enhanced in the similiar way.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>