[dualtor_io]: Enable autorestart for BGP tests (#4277)
What is the motivation for this PR?
During nightly test runs, container autorestart is disabled. This prevents the BGP container from exiting when the bgpd process is killed during this set of test cases, which prevents a switchover from occurring.

How did you do it?
Prior to the test, enable container autorestart for the BGP container. After the test is over, restore the previous autorestart state

How did you verify/test it?
Run the tests with container autorestart originally disabled. Verify that the tests are able to pass, and that autorestart for BGP is disabled after the test ends.

Signed-off-by: Lawrence Lee <lawlee@microsoft.com>