Use wait_until to check interface state in test_neighbor_mac_noptf to improve stability (#4291)
In setupDutConfig of this test, it will do config reload, and the wait time is 120s. Sometimes for some platforms such as sn3800, 120s is not enough for interface to be up. So use wait_until to check the interface is up or not, it will make the test more stable.

What is the motivation for this PR?
Use wait_until to check the interface is up or not, and make the test more stability.

How did you do it?
Use wait_until to check the interface is up or not.

How did you verify/test it?
run the test : test_neighbor_mac_noptf.py