[multi-asic][vs]: Use the right KVM platform string for multi-asic (#4335)
What is the motivation for this PR?
Error is seen when doing add-topo to bring up a multi-asic testbed.
./testbed-cli.sh -t vtestbed.csv -m veos_vtb add-topo vms-kvm-four-asic-t1-lag password.txt
..
TASK [find interface name mapping and individual interface speed if defined from dut] ***
..
 "msg": "failed to find the correct port config for msft_four_asic_vslist index out of range"
Reason:
Azure/sonic-buildimage#8269 added new multi-asic KVM platforms, for 4-asic and 6-asic VS.
When add-topo is done, port_alias.py runs on localhost to get the right front-panel interfaces so that VS can be brought up with the right number of front-panel interfaces. With the above PR, new platform string was added due to which port-config.ini file was under a new KVM platform directory, due to which the port config file was not being retrieved correctly.

How did you do it?
Based on number of asics, use the right KVM platform string so that the right port-config file can be retrieved when bringing up VS testbed.

How did you verify/test it?
Was able to bring up VS testbed without the above error.