Specify neighbor metadata for tgen topology (#4352)
Fixes #4223
Signed-off-by: Neetha John <nejo@microsoft.com>

What is the motivation for this PR?
On running RDMA tests, RX drops were seen on lag member interface. This was due to no pg profile attached to this port since the cable len was set to 0m. The cable len table is populated based on DEVICE_NEIGHBOR and DEVICE_NEIGHBOR_METADATA. DEVICE_NEIGHBOR_METADATA table is currently missing for tgen topology. This PR makes the necessary enhancement needed so that this table gets generated during minigraph deploy.

How did you do it?
Changes done so that the 'Devices' section of the minigraph will get populated with the neighbor info during rendering minigraph_png.j2 for tgen topology

How did you verify/test it?
Ran deploy-mg on tgen topology and ensured that the necessary table was generated
Ran deploy-mg on backend T0 topology to ensure that the new changes didn't cause any regression