[dual-tor] Correct the expectation num of pkts in IPinIP tunnel (#4288)
* [dual-tor] Correct the expectation num of pkts in IPinIP tunnel

What is the motivation for this PR?
As the outer header is the same in IPinIP tunnel between lower tor and upper tor, so within a single port-channel, all the pkts should send to the same interface. The other one should not receive the pkts. But in the current code, it is supposed the pkts will be received by all the interfaces evenly.

How did you do it?
Correct the expectation num of pkts according to the number of port-channels.

How did you verify/test it?
Run dualtor/test_orchagent_active_tor_downstream.py test case.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>

* [dual-tor] Correct the code according to the PR comments

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>