Fix issue of restart-ptf for dualtor testbed
There is an issue with running testbed-cli.sh restart-ptf on dualtor testbed. It failed at the
tasks of stopping and starting mux simulator.

The issue was introduced by https://github.com/Azure/sonic-mgmt/pull/4259. PR #4259 added
dependency of variable 'testbed_name' in playbook. The variable was not passed to playbook
in the testbed-cli.sh script.

The fix is to pass variable 'testbed_name' to playbook 'testbed_renumber_vm_topology.yml' in
the testbed-cli.sh tool. This PR also removed a legacy condition checking which failed in
control_mux_simulator.py.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>