[pfc] Fix issue in unknown mac test case
What is the motivation for this PR?
Currently, after the IP of a random interface in ptf will be clear at the end of each case. That means this interface can respond to the ARP request during the test period. If that happens, the DUT's fdb table is learned.

How did you do it?
Clean the IP at the beginning of each test.

How did you verify/test it?
Run pfc/test_unknown_mac.py

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>