[power reboot] improve robustness of power reboot test
When power reboot fails, it could leave the DUT powering back up
and taking a few minutes to become healthy again. Meanwhile
subsequent test will fail because DUT is unreachable or not
healthy.

To avoid the confusion, explicitly turn on all powers in case
of test failure and wait for DUT to become healthy again before
failing.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>