[test_configurable_drop_counters] Add static fdb entry for test_neighbor_link_down test case (#4463)
What is the motivation for this PR?
In test_neighbor_link_down, it will link down a specific egress interface. After that, swss will flush the fdb entries which are bonding to that egress interface. So the packets from the ingress may not able to find the output interface as the fdb entry is cleared. So the pkts will be dropped at the rx side, and the drop counter that we configured with L3_EGRESS_LINK_DOWN type would not work as expected.

How did you do it?
Add a static fdb entry for the egress interface as the fdb flush would just affect the dynamic entries. So in this way, the pkts could always find out the egress interface even the interface is down.

How did you verify/test it?
Run test_neighbor_link_down test case.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>