Set the maximum-paths to exabgp to 0 on v/ceos VM's when advertising more than 12K routes
In a T2 topology, when advertising more than 12K routes over the exabgp session to a ceos docker, the exabgp
BGP peership is not established. ceos complains about MaxPath

  ARISTA01T3(config)#show ip bgp summary
  BGP summary information for VRF default
  Router identifier 100.1.0.1, local AS number 65200
  Neighbor Status Codes: m - Under maintenance
  Description              Neighbor         V  AS           MsgRcvd   MsgSent  InQ OutQ  Up/Down State   PfxRcd PfxAcc
    65100                    10.0.0.0         4  65100          43194     11116    0    0 00:03:51 Estab   123    123
    exabgp_v4                10.10.246.254    4  65200          24116      2751    0    0 00:03:19 Idle(MaxPath)

To avoid this issue, we need to set 'maximum-routes' for exabgp neighbors to 0.
This was already done for neighbor to the DUT.