Fix SFP test failures on Arista 7050QX-32S platform (#4464)
Platforms like 7050QX-32S doesn't map all of their physical ports (1 to 36) to the logical ports because of which there could be unused indices in sequence of port indices. The current test assumes all physical port indices are mapped to logical ports. The SFP objects are stored sequentially(increasing order of physical port indices) in the Chassis object so there is no need to do 1:1 mapping as long as the SFP object is in the list of Chassis's SFP objects. Also, platform.json stores the list of SFP port names(Eg SFP1, SFP2, QSFP4, QSFP5...and so on in the increasing order of physical port indices). So, its sufficient to check if the SFP name is contained in the platform.json

Fixes #4444
Signed-off-by: Prince George <prgeor@microsoft.com>

How did you verify/test it?
Run test_sfp.py and test_chassis.py and ensure no failures