Update the docker-ptf and the docker-sonic-mgmt docker's download link (#4462)
Update the download link of docker-sonic-mgmt from Jenkins to Azurewebsites.
Add the download link of docker-ptf

What is the motivation for this PR?
The old Jenkins website is no longer for use. The original document is not friendly to beginners.

How did you do it?

How did you verify/test it?
Download docker-ptf and docker-sonic-mgmt from the links.