[qos] Disable IPv6 during the HdrmPoolSizeTest case
What is the motivation for this PR?
In HdrmPoolSizeTest case, it will send pkts and check the tx_drop counter which should not increase. But other services will send ICMPv6 neighbor advertisement pkts out for this egress interface. That will tx_drop counter increase.

How did you do it?
Disable IPv6 before the test to avoid the redundant NDP pkts getting involved in this case. Enable it after the test case is finished.

How did you verify/test it?
Run the testQosSaiHeadroomPoolSize and testQosSaiHeadroomPoolWatermark.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>