Merged PR 211: [DO NOT MERGE YET] Move testbeds on server_17 to server_24
Move testbeds on server_17 to server_24

Most lab test servers are running Ubuntu 18.04 which is a little bit old.
Server_24 is running Ubuntu 20.04 which is the current Ubuntu LTS
release. Eventually we'll need to upgrade all test servers to Ubuntu 20.04.
This change moved testbeds on server_17 to server_24 to see how
nightly test goes on Ubuntu 20.04. If everything looks good, we can
gradually upgrade all the test servers to Ubuntu 20.04 in the future.

To avoid disruption of nightly tests, need to merge this PR using below steps:
* Run remove-topo for testbeds on server_17.
* Merge this PR
* Run add-topo for new testbeds moved to server_24
* Rename the affected pipelines

Need to do the above steps on a Saturday evening when there is no nightly tests running

Extra steps to do:
* Update the Jenkins job for 3800-azd which is still used by Kavin.
* Send out email notifying this change.
* Update the testbed booking system.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>