Merged PR 232: Remove the hard coded server names of the recover_server pipeline
Remove the hard coded server names of the recover_server pipeline

The recover_server pipeline hard coded the list of server names in the its
yaml file. After a new server is onboarded, the pipeline yaml file needs to
be updated to include the new server name. This is inconvienent.

This change removed the hard coded server name list in the  recover
server pipeline. With this change, pipeline users need to explicitly
specify the server name instead of selecting server name from the hard
coded list.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>