Merged PR 240: Run a private Bullseye image as part of nightly test
Run a private Bullseye image as part of nightly test

For a Celestica DX010 device, run a private Bullseye image instead of
the current master image for the nightly test. This is so that there'll
be some data on how the Bullseye image is doing, since this image will
be merged into master in 2 weeks.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>