Improve Ptfadapter to randomize server port (#4510)
Description of PR
If a PTF test script with PTF adapter object paused at PDB interactive shell, the PTF adapter will retry connecting to the PTF nn agent running in PTF. While the PTF nn agent can only serve 1 connection, this would block other PTF tests from running.
The idea is to randomize the PTF nn agent listening port when the PTF nn agent is restarted. Then the zombie PTF adapter won't be able to occupy the server port.
In case the listening port is the same after restart and got occupied by zombie PTF adapter again, we should give explicit error message after PTF nn agent is restarted.

What is the motivation for this PR?
The idea is to randomize the PTF nn agent listening port when the PTF nn agent is restarted. Then the zombie PTF adapter won't be able to occupy the server port.
In case the listening port is the same after restart and got occupied by zombie PTF adapter again, we should give explicit error message after PTF nn agent is restarted.

How did you do it?
First, the previous ptf nn port is given a default port, I change this default port to a range of 100 consecutive ports.
Second, when starting the ptf nn agent in the docker, it selects a port from the above consecutive ports, and if the port is occupied by other threads, it will retry to choose another port. When it retries three times and can still not select a available port, the 'assert' will give an error. Otherwise, we start the ptf nn agent successfully.

How did you verify/test it?
I use the test case bgp/test_bgpmon.py to test the ptfadapter fixture. Including the port is available and unavailable. But the nbr_ptfadapter fixture doesn't have any test cases, so this fixture is not tested.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>