collect_dut_info to check if we have front-end and/or back-end asics (#4070)
What is the motivation for this PR?
PR #3696 added container names in meta data file for frontend and backend ASIC on which autorestart test will be performed for multi-asic platform. This assumes that on a multi-asic platform, there are always front-end and back-end asics.

However, on a VoQ chassis

linecards don't have back-end asics.
supervisor card has niether front-end nor back-end asics
So, need to check if we have front-end or back-end asics before trying to add them to asic_services.

How did you do it?
If we have front-end or back-end asics, (len != 0), only then pick up randomly and add it to asic_services.

How did you verify/test it?
Tested on multi-asic linecard and supervisor card in a VoQ chassis.