Set critical service list for a supervisor card based on number of SFMs in a VoQ chassis (#3813)
* Set critical service list for a supervisor card based on number of SFMs in a VoQ chassis

On a supervisor card, we have the global dockers pmon, lldp, database, and snmp.
For each Switch Fabric (SF) card in a VoQ Chassis, there could be n number of asics per card - where n is vendor specific.
However, whether the asic instance is created (syncd/swss etc. dockers) for a SF card would be up only if that SF is present/online.
For example, if a chassis supports max 8 SFM's and 2 asics per SFM, then we would expect 16 asic instances.
But if the chassis only 2 out of max 8 SFM's, then we would only have 4 asic instances.

So, we need to dynamically figure out which asic instances would be present based on the number of SFM's that are present and also in which slot they are present.
This data is derived by querying the chassis db using the following command
    redis-cli -h 10.0.5.16 -p 6380 -n 13  keys "CHASSIS_ASIC_TABLE|asic*"

The output of this gives a line per asic that is created/active.

Another difference on asics on supervisor card is that there would be no 'bgp' and 'teamd' dockers created as there are no BGP sessions and ports on a SFM.
For this we modified sonic_asic.get_critical_services to use _DEFAULT_SUPERVISOR_ASIC_SERVICES instead of _DEFAULT_ASIC_SERVICES if we are a supervisor card.

* Disable syslog rate limit to only disable on containers that are in the critical services list

On a supervisor card in a VoQ chassis, the number of syncd/teamd/swss containers
depends on the number of Switch Fabric cards present. The dockers that are created
are present in the critical_services_list. So, check if the docker (service) being
disabled is in the critical_service_list before trying to disable syslog rate
limit on that docker (service).

* Using chassis_db_address instead of hard-coded 10.0.5.16 to get data from CHASSIS_ASIC_DB (#3813 - sanmalho-git:sup_crit)

* Use sonic-db-cli instead of redis-cli