Merged PR 257: Add pipeline yaml file for weekend testbeds deep clean
Occasionally some unhealthy testbeds can break the whole test server and affect other testbeds on the same server. This change added a pipeline yaml for weekend testbeds deep clean. The pipeline will firstly try to lock all testbeds on a test server. If all testbeds are locked, then it will try to run a cleanup script to remove all the testbed related configurations on the test server. Patch the test server with latest upgrades. Reboot the server when necessary. Then recover all the testbeds and release them eventually.

This operation is too disruptive. To avoid negative impact to people's work, the pipeline will skip the server if any of the testbed on it is already locked by anyone.

We have updated all other pipelines to skip nightly tests on Saturday. This pipeline will only be executed every Saturday nightly to avoid conflicting with nightly tests.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>