Merged PR 236: Add pipeline yaml file for general nightly testing
This change added a pipeline yaml file for general nightly testing on any
testbed. This yaml file exposes all the parameters available in the template.
No scheduler is configured. Only manual trigger is supported.

To run this pipeline using other sonic-mgmt-int branch, please ensure that
this yaml file is available in that branch.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>