[Mellanox] [QoS/Dynamic_buffer] get lossy pg size via number lanes of a port instead of port speed (#4475)
On the Mellanox platform, ports with 8 lanes have different lossy pg size with ports have 4 or fewer lanes.
The previous assumption is that for 8 lanes port the speed will always be 400G, so the key for the lossy PG size value is port speed. But this is not true, it can also be configured to 100G or other speeds. In these cases will not be able to get the correct lossy PG size. So use the number of lanes as the key for the lossy PG size is proper, it's stable in the lifetime of a port.

- How did you do it?
Use the number of lanes as the key for the lossy PG size, instead of using port speed.

- How did you verify/test it?
run test_buffer:test_port_admin_down on a testbed with 8 lanes port and speed was set to none 400G.

- Any platform specific information?
Mellanox platform only.