[T2 topology] Fix ASN numbers for T1s and T3s (#4497)
In a typical DC topology a T2 router connects to a T1 router from a different podset. Each podset is assigned a unique ASN. And in upstream direction T2 connects to T3s belonging to a single IDF (or plane) that have the same ASN.

Also updated the AS path for routes advertised in T3 VMs.