Merged PR 266: Update lab SONiC DUTs to user server_16 as syslog server
The existing syslog servers are not really working for some reasons. We need a way
to archive syslog of SONiC DUTs for later troubleshooting in case the DUTs are
fresh re-installed and all the logs are lost.

We can create a new Azure VM like trusty8 to be used as syslog server. However,
there would be more work for maintaining who has access to that server.

The better idea is to use a lab server without heavy work load as syslog server.

Syslog server has been configured on lab server_16. This change updated the the
syslog server of SONiC DUTs to IP address of server_16. After the SONiC DUTs
are redeployed with a new minigraph, then they will use server_16 as syslog server.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>